requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.0
alembic==1.13.2
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0